Web/02 - Generateur Vraiment Super Secure/source/pyproject.toml
[tool.poetry]
name = "chall-2"
version = "0.1.0"
description = ""
authors = ["tristanqtn <tristan.querton@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyjwt = "^2.10.1"
flask = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
